lambdas/thumbnail-generator/requirements.txt
boto3==1.35.77
Pillow==11.0.0
